requirements.txt
absl-py
amqp==2.2.2
aniso8601==7.0.0
appdirs==1.4.3
appnope==0.1.0
APScheduler==3.4.0
arrow
astor==0.8.0
astroid==1.6.1
atomicwrites==1.3.0
attrs==19.1.0
autopep8
avinit
Babel==2.5.1
babis==0.2.2
bcrypt
beautifulsoup4==4.8.1
billiard==3.5.0.3
bitly-api-py3
bleach==3.1.0
boto==2.49.0
cached-property==1.3.0
cachetools==3.1.1
captcha==0.3
celery==4.1.0
celery-singleton==0.3.1
certifi==2019.9.11
cffi
chardet==3.0.4
chartkick==0.5.0
Click==7.0
cloudflare==2.3.0
colorama==0.4.1
configparser==3.5.0
confluent_kafka==1.8.2
contextlib2==0.5.4
cssmin==0.2.0
Cython==0.29.4
dateparser==0.5.0
decorator==4.0.10
defusedxml==0.5.0
dnspython==1.15.0
docopt==0.6.2
docutils==0.14
dominate==2.2.1
elastic-apm==5.10.1
elasticsearch==7.0.1
email-validator==1.1.2
enum34==1.1.6
et-xmlfile==1.0.1
facepy==1.0.12
fbmq==2.3.2
flake8==3.5.0
Flask==1.1.2
Flask-Assets==0.12
Flask-Babel==0.11.2
Flask-BasicAuth==0.2.0
Flask-Bootstrap==3.3.7.0
Flask-CacheBuster==1.0.0
Flask-Caching==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.8
Flask-DebugToolbar==0.11.0
Flask-Excel==0.0.7
Flask-GraphQL==2.0.0
Flask-HTMLmin==1.5.0
Flask-Limiter==0.9.3
flask-mongoengine
Flask-OAuthlib==0.9.3
flask-paginate==0.4.5
Flask-RESTful==0.3.5
flask-restplus==0.9.2
Flask-SeaSurf==0.2.2
Flask-Session
flask-session-captcha==1.2.0
Flask-SQLAlchemy==2.5.1
flask-talisman==0.7.0
Flask-WTF==0.14.2
flower==0.9.3
funcsigs==1.0.2
# functools32==3.2.3.post2
furl==2.0.0
future==0.16.0
futures==3.1.1
geopy==1.11.0
#gevent
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
graphene==2.1.8
graphene-mongo==0.2.8
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
greenlet
gunicorn==19.9.0
hashids==1.2.0
htmlmin==0.1.12
httplib2==0.19.0
idna==2.8
importlib-metadata==2.1.1
iniparse==0.4
ipaddress==1.0.16
IPy==0.75
ipython==5.1.0
ipython-genutils==0.1.0
iso8601==0.1.12
isodate
isort==4.3.4
itsdangerous==1.1.0
jdatetime==1.8.1
jdcal==1.3
jedi==0.10.2
Jinja2==2.10.3
jsmin
json-rpc==1.10.3
jsonlines==1.2.0
jsonschema==2.5.1
kombu==4.1.0
lazy-object-proxy==1.3.1
limits==1.2.0
lml==0.0.1
lockfile==0.12.2
luigi==2.8.8
lxml==4.8.0
mailchimp3==3.0.14
MarkupSafe==1.1.1
mccabe==0.6.1
meld3==1.0.2
mercurial
minio==4.0.6
mongobackup==0.2.4
mongoengine
mongomock==3.14.0
netifaces==0.10.6
numpy==1.19.5
O365
oauthlib==2.0.1
openpyxl==2.4.8
orderedmultidict==1.0
packaging==20.9
#pandas==0.24.2
pathlib2==2.1.0
pep8==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
pika==1.1.0
#Pillow==6.2.0
pipreqs==0.4.9
pluggy==0.4.0
ply==3.4
promise==2.2.1
prompt-toolkit==1.0.9
protobuf==3.15.3
psycopg2-binary
ptftpd==1.2
ptyprocess==0.5.1
py==1.4.33
pyasn1
pyasn1-modules
pycodestyle==2.3.1
pycparser==2.17
pyexcel==0.5.3
pyexcel-io==0.4.2
pyexcel-webio==0.1.2
pyexcel-xlsx==0.4.0
pyfacebook==0.3
pyflakes==1.6.0
Pygments==2.1.3
pylint==1.8.2
pymongo==3.13.0
pymssql
pyotp==2.2.6
pyparsing==2.4.7
PyQRCode==1.2.1
python-daemon==2.1.2
python-dateutil==2.8.0
python-http-client==3.3.2
python-language-server==0.5.0
python-memcached==1.58
python-slugify==1.2.0
pytz==2016.7
PyYAML==5.1.1
raven==5.31.0
redis==2.10.6
regex==2016.10.22
requests==2.22.0
requests-oauthlib==0.7.0
requests-toolbelt==0.8.0
rq==0.6.0
rq-dashboard==0.3.7
rsa==3.4.2
Rx==1.6.1
schedule==0.4.3
sendgrid==6.6.0
sentinels==1.0.0
sentry-sdk==0.20.3
# simple-pbkdf2==1.0
simplegeneric==0.8.1
simplejson==3.10.0
singledispatch==3.4.0.3
six
slacker==0.9.60
soupsieve==1.9.4
SQLAlchemy==1.4.15
starkbank-ecdsa==1.1.0
supervisor
texttable==0.9.1
tinify==1.5.1
tornado==4.5.3
tox==2.7.0
traitlets==4.3.1
typing==3.7.4.1
tzlocal==1.3
ua-parser==0.7.1
umalqurra==0.2
# umysql==2.61
unicode-slugify==0.1.3
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.6
user-agents==1.0.1
validate-email==1.3
validators==0.11.0
validictory==1.1.1
vine==1.1.4
virtualenv==15.1.0
visitor==0.1.3
wcwidth==0.1.7
webassets==0.12.0
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.10.11
WTForms==2.2.1
xlrd==1.2.0
XlsxWriter==1.0.0
yapf==0.13.2
yarg==0.1.9
zalo-python-sdk==1.15
zeep==2.2.0
zipp==1.2.0
cherrypy
kafka-python
PyMySQL
waitress
